Move SearchTextBox to better namespace
Is being used in multiple places now.
